Restore lib9 to its rightful glory: fix include/lib9.h, lib9 sources, CI packages
Agent-Logs-Url: https://github.com/cogpy/cogfernos/sessions/a0e83e80-6b49-4d76-8026-ea9f205d7adc

+++ lib9/create.c
@@ -1,10 +1,9 @@
-#include <sys/types.h>
-#include <sys/stat.h>
-#include <fcntl.h>
 #include "lib9.h"
+#include <sys/types.h>
+#include <fcntl.h>
 
 int
-create(char *f, int mode, ulong perm)
+create(char *f, int mode, int perm)
 {
 	int m;
